requirements.txt
networkx>=3.0
pandas>=2.0
numpy>=1.24
click>=8.0
pyyaml>=6.0
pulp>=2.7
scipy>=1.11
pytest>=7.0
rich>=13.0
requests>=2.31
